backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-browser-reload==1.18.0",
    "django-compressor==4.5.1",
    "django-crispy-forms==2.4",
    "djangorestframework==3.16.0",
    "markdown==3.8",
    "psycopg==3.2.8",
    "python-decouple==3.8",
]
